refactoring

+++ src/prerequisites/views/prerequisite.py
@@ -1,3 +1,4 @@
+from django.http import Http404
 from django.shortcuts import render, get_object_or_404, redirect
 from django.contrib.auth.decorators import login_required
 from django.utils.translation import ugettext_lazy as _
@@ -7,27 +8,36 @@ from registration.decorators import archived_not_available
 from registration.models import Event, Helper
 from registration.views.utils import nopermission
 
-from .forms import PrerequisiteForm, PrerequisiteDeleteForm
-from .models import Prerequisite
+from .utils import notactive
+
+from ..forms import PrerequisiteForm, PrerequisiteDeleteForm
+from ..models import Prerequisite
 
 import logging
 logger = logging.getLogger("helfertool")
+
 
 @login_required
 @archived_not_available
 def edit_prerequisite(request, event_url_name, prerequisite_pk=None):
     event = get_object_or_404(Event, url_name=event_url_name)
 
-    # permission
+    # check if feature is active
+    if not event.prerequisites:
+        return notactive(request)
+
+    # check permission
     if not event.is_admin(request.user):
         return nopermission(request)
 
     # get prerequisite, if available
     prerequisite = None
-    if prerequisite:
-        prerequisite = get_object_or_404(prerequisite, pk=prerequisite_pk)
+    if prerequisite_pk:
+        prerequisite = get_object_or_404(Prerequisite, pk=prerequisite_pk)
+        if prerequisite.event != event:
+            raise Http404
 
-    # form
+    # handle form
     form = PrerequisiteForm(request.POST or None, instance=prerequisite, event=event)
 
     if form.is_valid():
@@ -40,7 +50,8 @@ def edit_prerequisite(request, event_url_name, prerequisite_pk=None):
         logger.info(log_msg, extra={
             'user': request.user,
             'event': event,
-            'prerequisite': prerequisite,
+            'prerequisite': prerequisite.name,
+            'prerequisite_pk': prerequisite.pk,
         })
 
         # redirect to prerequisite overview
@@ -56,7 +67,16 @@ def edit_prerequisite(request, event_url_name, prerequisite_pk=None):
 @login_required
 def view_prerequisites(request, event_url_name):
     event = get_object_or_404(Event, url_name=event_url_name)
-    prerequisites = Prerequisite.objects.all()
+
+    # check if feature is active
+    if not event.prerequisites:
+        return notactive(request)
+
+    # check permission
+    if not event.is_admin(request.user):
+        return nopermission(request)
+    
+    prerequisites = Prerequisite.objects.filter(event=event)
 
     context = {'event': event,
                'prerequisites': prerequisites}
@@ -67,25 +87,32 @@ def view_prerequisites(request, event_url_name):
 @archived_not_available
 def delete_prerequisite(request, event_url_name, prerequisite_pk):
     event = get_object_or_404(Event, url_name=event_url_name)
-    prerequisite = get_object_or_404(Prerequisite, pk=prerequisite_pk)
+
+    # check if feature is active
+    if not event.prerequisites:
+        return notactive(request)
 
     # check permission
     if not event.is_admin(request.user):
         return nopermission(request)
+
+    prerequisite = get_object_or_404(Prerequisite, pk=prerequisite_pk)
+
+    if prerequisite.event != event:
+        raise Http404
 
     # form
     form = PrerequisiteDeleteForm(request.POST or None, instance=prerequisite)
 
     if form.is_valid():
         form.delete()
-        messages.success(request, _("Prerequisite deleted: %(name)s") %
-                         {'name': prerequisite.name})
+        messages.success(request, _("Prerequisite deleted: %(name)s") % {'name': prerequisite.name})
 
         logger.info("prerequisite deleted", extra={
             'user': request.user,
             'event': event,
-            'prerequisite': prerequisite,
-            'prerequisite_pk': prerequisite_pk,
+            'prerequisite': prerequisite.name,
+            'prerequisite_pk': prerequisite.pk,
         })
 
         # redirect to prerequisite overview
@@ -96,21 +123,3 @@ def delete_prerequisite(request, event_url_name, prerequisite_pk):
                'prerequisite': prerequisite,
                'form': form}
     return render(request, 'prerequisites/delete_prerequisite.html', context)
-
-
-@login_required
-def view_helpers_prerequisite(request, event_url_name, prerequisite_pk):
-
-    event = get_object_or_404(Event, url_name=event_url_name)
-    prerequisite = get_object_or_404(Prerequisite, pk=prerequisite_pk)
-
-    # find all helpers that need this prerequisite
-    helpers = Helper.objects.all()\
-        .filter(shifts__job__prerequisites=prerequisite)
-
-    # render page
-    context = {'event': event,
-               'prerequisite': prerequisite,
-               'helpers': helpers}
-    return render(request, 'prerequisites/view_helpers_prerequisite.html', context)
-
